Code/requirements.txt
wheel
pywebview
youtube-dl
requests